% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_mcmc.R
\name{da.mgp}
\alias{da.mgp}
\title{Data augmentation with a pseudo-marginal Markov chain Monte Carlo algorithm for multivariate generalized Pareto models}
\usage{
da.mgp(dat, mthresh, thresh, lambdau = 1, model = c("br", "xstud"),
  coord, start, numiter = 40000L, burnin = 5000L, thin = 1L,
  verbose = 50L, filename, keepburnin = TRUE, geoaniso = TRUE,
  blockupsize = ncol(dat), likt = c("mgp", "pois", "binom"),
  saveinterm = 500L, ...)
}
\arguments{
\item{dat}{n by D matrix of observations}

\item{mthresh}{vector of marginal thresholds under which data are censored}

\item{thresh}{functional max threshold determining the risk region}

\item{lambdau}{probability of exceedance of the threshold for censored observations}

\item{model}{dependence model, either of \code{"br"}, \code{"xstud"} or \code{"lgm"}, in which case the model uses the independence likelihood with generalized Pareto margins}

\item{coord}{matrix of coordinates, with longitude and latitude in the first two columns and additional covariates for the latent Gaussian model}

\item{start}{named list with starting values for the parameters, with arguments:
\itemize{
\item\code{scale}: a \code{D} vector of scale parameters, strictly positive.
\item\code{shape}: a scale containing the shape and satisfying support constraints for all sites.
\item\code{marg.pcov}: initial proposal covariance for the marginal parameters (scale and shape).
\item\code{dep}: initial values for the dependence function.
\item\code{dep}: dependence function, with distance matrix as first argument and \code{dep} as second argument.
\item\code{aniso}: initial values for the anisotropy parameters, scale and angle.
\item\code{df}: degrees of freedom (if \code{model == "xstud"}).
\item\code{dep.lb}: lower bounds for the dependence parameters
\item\code{dep.ub}: upper bounds for the dependence parameters
}
If any of \code{scale}, \code{shape} or \code{marg.pcov} are missing, the function will attempt to find starting values.}

\item{numiter}{number of iterations to be returned}

\item{burnin}{number of initial parameters for adaptation and discarded values.}

\item{thin}{thining parameter; only every \code{thin} iteration is saved}

\item{verbose}{report current values via print every \code{verbose} iterations.}

\item{filename}{name of file for save.}

\item{keepburnin}{logical; should initial runs during \code{burnin} be kept for diagnostic. Default to \code{TRUE}.}

\item{geoaniso}{logical; should geometric anisotropy be included? Default to \code{TRUE}.}

\item{blockupsize}{size of block for updates of the scale parameter; \code{ncol(dat)} yields individual updates}

\item{likt}{string indicating the type of likelihood, with an additional contribution for the non-exceeding components: one of  \code{"mgp"}, \code{"binom"} and \code{"pois"}.}

\item{saveinterm}{integer indicating when to save results. Default to \code{500L}.}

\item{...}{Arguments passed on to \code{clikmgp}
\describe{
  \item{mthresh}{vector of individuals thresholds under which observations are censored}
  \item{dat}{matrix of observations}
  \item{thresh}{functional threshold for the maximum}
  \item{loc}{vector of location parameter for the marginal generalized Pareto distribution}
  \item{scale}{vector of scale parameter for the marginal generalized Pareto distribution}
  \item{shape}{vector of shape parameter for the marginal generalized Pareto distribution}
  \item{par}{list of parameters: \code{alpha} for the logistic model, \code{Lambda} for the Brown--Resnick model or else \code{Sigma} and \code{df} for the extremal Student.}
  \item{model}{string indicating the model family, one of \code{"log"}, \code{"br"} or \code{"xstud"}}
  \item{likt}{string indicating the type of likelihood, with an additional contribution for the non-exceeding components: one of  \code{"mgp"}, \code{"binom"} and \code{"pois"}.}
  \item{lambdau}{vector of marginal rate of marginal threshold exceedance.}
}}
}
\value{
a list with \code{res} containing the results of the chain
}
\description{
Data augmentation with a pseudo-marginal Markov chain Monte Carlo algorithm for multivariate generalized Pareto models
}
\keyword{internal}
