% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condsimu.R
\name{dcondmvtnorm}
\alias{dcondmvtnorm}
\title{Conditional density of Gaussian or Student subvectors}
\usage{
dcondmvtnorm(x, ind, lbound = rep(-Inf, length(ind)), ubound = rep(Inf,
  length(ind)), mu, sigma, df = NULL, model = c("norm", "stud"),
  n = 500, log = FALSE)
}
\arguments{
\item{x}{a \code{p} vector with the values of process at the remaining coordinates}

\item{ind}{a \code{d} vector of integer indices in \eqn{\{1, \ldots, d+p\}} for which to compute the conditional density}

\item{lbound}{\code{d} vector of lower bounds}

\item{ubound}{\code{d} vector of upper bounds for truncated}

\item{mu}{\code{d+p} vector of location parameters}

\item{sigma}{\code{d+p} by \code{d+p} scale matrix}

\item{df}{degrees of freedom of the \code{d+p} dimensional Student process}

\item{model}{string indicating family, either \code{norm} for Gaussian or \code{stud} for Student-t}

\item{n}{sample size for simulations. Default to 500.}

\item{log}{logical; should log probability be returned? Default to \code{FALSE}.}
}
\value{
the conditional (log) density of the vector \code{x[ind]}
}
\description{
The function computes the conditional density of (truncated) Gaussian or Student components corresponding to indices \code{ind},
 given the values at the remaining index.
The location vector \code{mu} and the scale matrix \code{sigma} are those of the \eqn{d+p} vector.
The routine relies on the CDF approximation based on minimax exponential tilting implemented in the \code{TruncatedNormal} package.
}
