% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{dinvgamma}
\alias{dinvgamma}
\title{Inverse gamma density, parametrized in terms of shape and rate}
\usage{
dinvgamma(x, alpha, beta, log = TRUE)
}
\arguments{
\item{x}{vector of observations}

\item{log}{logical; should the log-density be returned?}

\item{shape}{shape parameter, strictly positive}

\item{rate}{rate parameter, strictly positive}
}
\description{
Inverse gamma density, parametrized in terms of shape and rate
}
