% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{dmvnorm.precis}
\alias{dmvnorm.precis}
\title{Multivariate normal density function parametrized in terms of a precision matrix}
\usage{
dmvnorm.precis(x, mean, precis, logd = FALSE)
}
\arguments{
\item{x}{matrix or vector of observations}

\item{mean}{mean vector}

\item{precis}{precision matrix}

\item{logd}{logical; whether log-density should be returned (default to \code{FALSE})}
}
\value{
density or log-density of the \code{nrow(x)} sample
}
\description{
This function returns the log-density for a multivariate Gaussian distribution.
}
