% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Multivariate normal density function}
\usage{
dmvnorm(x, mean, sigma, logd = FALSE)
}
\arguments{
\item{x}{matrix or vector of observations}

\item{mean}{mean vector}

\item{sigma}{positive definite covariance matrix}

\item{logd}{logical; whether log-density should be returned (default to \code{FALSE})}
}
\value{
density or log-density of the \code{nrow(x)} sample
}
\description{
This function returns the log-density for a multivariate Gaussian distribution.
}
