% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{dmvstud}
\alias{dmvstud}
\title{Density of the multivariate Student density}
\usage{
dmvstud(x, mu = rep(0, p), sigma = diag(p), df = 1, logd = TRUE)
}
\arguments{
\item{x}{vector or matrix of observations}

\item{mu}{centrality parameter}

\item{sigma}{covariance matrix}

\item{df}{degrees of freedom parameter, default to 1 (Cauchy distribution)}

\item{logd}{logical; should log-density be returned? default to \code{TRUE}}
}
\description{
The function is adapted from \code{mvtnorm} package and computes the (log)-density
of the whole sample.
}
\keyword{internal}
