% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.dmvnorm_arma}
\alias{.dmvnorm_arma}
\title{Multivariate normal density function}
\usage{
.dmvnorm_arma(x, mean, sigma, logd = FALSE)
}
\arguments{
\item{x}{matrix of observations}

\item{mean}{mean vector}

\item{sigma}{positive definite covariance matrix}

\item{logd}{logical; whether log-density should be returned (default to \code{FALSE})}
}
\value{
density or log-density of the \code{nrow(x)} sample
}
\description{
This function returns the log-density for a multivariate Gaussian distribution.
The data must be imputed as a matrix, using e.g., \code{as.matrix}, with each row
representing an observation.
}
\keyword{internal}
