% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{expit}
\alias{expit}
\title{Expit inverse link function for logistic regression}
\usage{
expit(x)
}
\arguments{
\item{x}{parameter}
}
\value{
probability
}
\description{
The expit function is \eqn{\exp(x)/(1+\exp(x))}.
}
\keyword{internal}
