% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremogram.R
\name{extremo}
\alias{extremo}
\title{Pairwise extremogram for max-risk functional}
\usage{
extremo(dat, margp, loc, scale = 1, rho = 0, plot = FALSE, ...)
}
\arguments{
\item{dat}{data matrix}

\item{margp}{marginal probability above which to threshold observations}

\item{loc}{matrix of locations (one site per row)}

\item{scale}{geometric anisotropy scale parameter}

\item{rho}{geometric anisotropy angle parameter}

\item{plot}{logical; should a graph of the pairwise estimates against distance? Default to \code{FALSE}}

\item{...}{additional arguments passed to plot}
}
\value{
an invisible matrix with pairwise estimates of chi along with distance (unsorted)
}
\description{
The function computes the pairwise \eqn{chi} estimates and plots them as a function of the distance between sites.
}
