% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute}
\alias{impute}
\title{Data augmentation for the Brown-Resnick process}
\usage{
impute(dat, thresh, mthresh, loc = 1, scale = 1, shape = 1,
  lambdau = 1, riskr = c("max", "sum"), par, map = FALSE, ...)
}
\arguments{
\item{dat}{matrix of observations}

\item{mthresh}{vector of individuals mthresholds under which observations are censored}

\item{loc}{vector of location parameter for the marginal generalized Pareto distribution}

\item{scale}{vector of scale parameter for the marginal generalized Pareto distribution}

\item{shape}{vector of shape parameter for the marginal generalized Pareto distribution}

\item{lambdau}{vector of marginal rate of marginal threshold exceedance.}

\item{riskr}{string giving the risk region, either \code{max} or \code{sum}}

\item{par}{list of parameters, with \code{alpha} for the logistic model, \code{Lambda} for the Brown--Resnick model or else \code{Sigma} and \code{df} for the extremal Student.}

\item{map}{logical; should imputed data be returned on the unit Pareto scale? Default to \code{FALSE}}
}
\value{
a matrix of observations with imputed values in place of censored components.
}
\description{
This function returns a matrix of data on the data scale (if \code{map == TRUE}) or else unit Pareto observations.
Data that fall below a marginal threshold are censored and imputed according to the conditional distribution.
For \code{riskr == "sum"}, this require an accept-reject scheme to ensure the points with the simulated components
still lies in the risk region defined by the marginal and dependence parameters.
}
