% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{logit}
\alias{logit}
\title{Logistic link function}
\usage{
logit(x)
}
\arguments{
\item{x}{probability}
}
\value{
transformed probability
}
\description{
Canonical link function for logistic regression with binomial data,
\eqn{\log(x)-\log(1-x)}.
}
\keyword{internal}
