% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mh.fun}
\alias{mh.fun}
\title{Update function based on Metropolis-Hasting algorithm}
\usage{
mh.fun(cur, lb, ub, prior.fun, lik.fun, ll, ind, pmu, pcov, cond = TRUE,
  ...)
}
\arguments{
\item{cur}{current value of the vector of parameters of which one component is to be updated.}

\item{lb}{lower bounds for the parameter of interest. Default to \code{-Inf} if argument is missing.}

\item{ub}{upper bounds for the parameters of interest. Default to \code{Inf} if argument is missing.}

\item{prior.fun}{log prior function, a function of the parameter to update.}

\item{lik.fun}{log-likelihood function, a function of the parameter to  For dependence parameters, updates of the dependence parameters can be passed as attributes.}

\item{ll}{value of the log-likelihood at \code{cur}.}

\item{ind}{indices of \code{cur} to update.}

\item{pmu}{proposal mean; if missing, default to random walk.}

\item{pcov}{covariance matrix of the proposal for \code{cur}.}

\item{cond}{logical; should updates be made conditional on other values in \code{cur}. Default to \code{TRUE}.}

\item{...}{additional arguments passed to the function, currently ignored.}
}
\value{
a list with components
\itemize{
\item{\code{ll}}: value of the log-likelihood, with potential additional parameters passed as attributes;
\item{\code{cur}}: values of the parameters after update;
\item{\code{accept}}: logical indicating the proposal has been accepted (\code{TRUE}) or rejected (\code{FALSE}).
}
}
\description{
Proposals are made based on a (conditional) truncated Gaussian distribution at indices \code{ind}
given other components.
}
