% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condsimu.R
\name{pcondmvtnorm}
\alias{pcondmvtnorm}
\title{Conditional distribution of Gaussian or Student subvectors}
\usage{
pcondmvtnorm(n = 500, ind, x, lbound = rep(-Inf, length(ind)),
  ubound = rep(Inf, length(ind)), mu, sigma, df = NULL,
  model = c("norm", "stud"), log = FALSE)
}
\arguments{
\item{n}{sample size for simulations. Default to 500.}

\item{ind}{\code{d} vector of indices with integer entries in \eqn{\{1, \ldots, d+p\}} for which to compute the distribution function}

\item{x}{\code{d+p} vector}

\item{lbound}{\code{d} vector of lower bounds for truncation}

\item{ubound}{\code{d} vector of upper bounds for truncation}

\item{mu}{\code{d+p} vector of location parameters}

\item{sigma}{\code{d+p} by \code{d+p} scale matrix}

\item{df}{degrees of freedom of the \code{d+p} dimensional Student process}

\item{model}{string indicating family, either \code{norm} for Gaussian or \code{stud} for Student-t}

\item{log}{logical; should log probability be returned? Default to \code{FALSE}.}
}
\value{
conditional distribution function for the components \code{ind} at \code{x[ind]}
}
\description{
The function computes the conditional distribution of the sub-components in \code{ind} given the rest and
returns the distribution function of the truncated Gaussian or Student components.
The location vector \code{mu} and the scale matrix \code{sigma} are those of the \eqn{d+p} vector.
The routine relies on the CDF approximation based on minimax exponential tilting implemented in the \code{TruncatedNormal} package.
}
