% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{rmvnormprec}
\alias{rmvnormprec}
\title{Gaussian sampler using the Cholesky decomposition of the precision matrix}
\usage{
rmvnormprec(n, mu = rep(0, ncol(precis)), precis)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean vector}

\item{precis}{precision matrix}
}
\value{
sample of size \code{n} from the multivariate Gaussian distribution
}
\description{
Gaussian sampler using the Cholesky decomposition of the precision matrix
}
