% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{shape.lgm}
\alias{shape.lgm}
\title{Update for latent Gaussian model for the scale parameter of a generalized Pareto}
\usage{
shape.lgm(scale, shape, ldat, shape.mu = NULL, shape.precis = NULL,
  shape.tausq = NULL, mmax, lbound = -0.5, ubound = 0.5,
  discount = 1, maxstep = 0.1)
}
\arguments{
\item{scale}{vector of scale parameters for the generalized Pareto distribution}

\item{shape}{vector of shape parameters for the generalized Pareto distribution}

\item{ldat}{list with exceedances at each site}

\item{shape.mu}{mean of the Gaussian process for the shape parameters}

\item{shape.precis}{precision matrix of the Gaussian process corresponding to the inverse of the correlation matrix}

\item{shape.tausq}{variance of the Gaussian process for the shape parameters}

\item{mmax}{vector of maximum of each series in \code{ldat}}

\item{lbound}{lower bound if parameter is truncated below}

\item{ubound}{upper bound if parameter is truncated above}

\item{discount}{numeric giving the discount factor for the Newton  Default to 1.}

\item{maxstep}{maximum step size for the MCMC (proposal will be at most \code{maxstep} units away from the current value).}
}
\value{
a vector of scale parameters
}
\description{
The scale has a log-Gaussian prior with variance \code{lscale.tausq} and precision (the inverse of the correlation matrix)
given by \code{lscale.precis}.
}
