% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{wrapAng}
\alias{wrapAng}
\title{Wrap angular component in anisotropy}
\usage{
wrapAng(ang)
}
\arguments{
\item{ang}{angle}
}
\value{
angle between \eqn{(-\pi/2,\pi/2)}
}
\description{
Wrap angular component in anisotropy
}
\keyword{internal}
