% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{dinvgamma}
\alias{dinvgamma}
\title{Inverse gamma density, parametrized in terms of shape and rate}
\usage{
dinvgamma(x, shape, scale, log = TRUE)
}
\arguments{
\item{x}{vector of observations}

\item{shape}{shape parameter, strictly positive}

\item{scale}{rate parameter of the gamma distribution, strictly positive}

\item{log}{logical; should the log-density be returned?}
}
\description{
Inverse gamma density, parametrized in terms of shape and rate
}
